5.1. Add trip controller and move trip events rendering logics to it
